modules: create a separate module for flow information
